Some error when opening EditUserForm

+++ ChapeauxUI/DeleteAccountForm.cs
@@ -22,6 +22,7 @@ namespace ChapeauxUI
 
         private void butDeleteUserAccount_Click(object sender, EventArgs e)
         {
+            int effectedRows = 0;
             User_Service user_Service = new User_Service();
 
             bool emptyFields = false;
@@ -36,7 +37,7 @@ namespace ChapeauxUI
                 MessageBox.Show("Please Fill All Required Fields", "Required Fields are Empty", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
             else
-            {
+            {   //assigns the role
                 Role userRole;
                 if (rbManager.Checked)
                     userRole = Role.Manager;
@@ -55,19 +56,19 @@ namespace ChapeauxUI
                     LoginPassword = txtPassword.Text,
                     Role = userRole
                 };
-                
+
                 DialogResult dialog = MessageBox.Show("Are you sure?", "Delete Account", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
                 if (dialog == DialogResult.Yes)
                 {
-                    user_Service.RemoveUserAccount(deleteUser); //delete the User from the DB 
-
+                    if (effectedRows >= 0)
+                    {
+                        user_Service.RemoveUserAccount(deleteUser); //delete the User from the DB 
+                        MessageBox.Show("Account deletion successful!");
+                    }
                 }
                 else
-                {
                     MessageBox.Show("Could not delete the account!");
-                }
             }
-
         }
     }
 }
